backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
yt-dlp>=2024.0.0
mistralai>=0.1.0
faiss-cpu==1.7.4
numpy==1.26.3
pydantic==2.5.3
httpx==0.26.0
